+++ cpskin/core/viewlets/footer.py
@@ -1,7 +1,6 @@
 # -*- coding: utf-8 -*-
 from Acquisition import aq_inner
 from plone import api
-from plone.app.layout.viewlets.common import FooterViewlet
 from plone.app.layout.viewlets.common import ViewletBase
 from Products.CMFCore.utils import getToolByName
 from Products.Five.browser.pagetemplatefile import ViewPageTemplateFile
@@ -58,11 +57,3 @@ class CPSkinFooterSitemapViewlet(ViewletBase):
         if getattr(footer_static, 'text', None):
             text = footer_static.text.raw
         return text
-
-
-class CpskinFooterViewlet(FooterViewlet):
-
-    def is_gdpr(self):
-        return api.portal.get_registry_record(
-            'imio.gdpr.interfaces.IGDPRSettings.is_text_ready',
-            default=False)