+++ CHANGES.rst
@@ -4,7 +4,7 @@ Changelog
 0.13.1 (unreleased)
 -------------------
 
-- Add legal mentions in Sendinblue portlet : WEB-2886
+- Add legal mentions checkbox in Sendinblue portlet : WEB-2886
   [laulaz]
 
 - Fix viewlets.xml that didn't apply at all