+++ CHANGES.rst
@@ -9,6 +9,12 @@ Changelog
 
 - Add recaptcha on register form : WEB-2883
   [laulaz]
+  
+- Fix viewlets.xml that didn't apply at all
+  [laulaz]
+
+- Fix collapsable menu display for minisites : WEB-2885
+  [laulaz]
 
 
 0.13.0 (2019-01-08)